requirements.txt
streamlit
numpy
pillow
pandas
opencv-python-headless==4.10.0.84
torch
torchvision
ultralytics